add list of Gas-Station datas
